Src/Sat.Recruitment.Api/Middleware/ExceptionHandlerMiddleware.cs
using System;
using System.Net;
using Newtonsoft.Json;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Http;
using Sat.Recruitment.Service.Dto;
using System.Security.Authentication;
using System.ComponentModel.DataAnnotations;
using Sat.Recruitment.Service.CustomException;

namespace Sat.Recruitment.Api.Middleware
{
    public class ExceptionHandlerMiddleware
    {
        private readonly RequestDelegate _next;

        public ExceptionHandlerMiddleware(RequestDelegate next) => _next = next;

        public async Task Invoke(HttpContext context)
        {
            try
            {
                await _next(context);
            }
            catch (Exception ex)
            {
                var response = context.Response;
                response.ContentType = "application/json";
                response.StatusCode = (int)GetErrorCode(ex);

                var result = JsonConvert.SerializeObject(new ResultErrorDto
                {
                    IsSuccess = false,
                    Errors = ex.Message
                });
                await response.WriteAsync(result);
            }
        }

        private static HttpStatusCode GetErrorCode(Exception e)
        {
            switch (e)
            {
                case BaseException _:
                    return ((BaseException)e).StatusCode;
                case ValidationException _:
                    return HttpStatusCode.BadRequest;
                case FormatException _:
                    return HttpStatusCode.BadRequest;
                case AuthenticationException _:
                    return HttpStatusCode.Forbidden;
                case NotImplementedException _:
                    return HttpStatusCode.NotImplemented;
                default:
                    return HttpStatusCode.InternalServerError;
            }
        }
    }
}